Identity.getNextEdition(): Rename to getNextEditionToFetch(); document
To make it less likely that someone uses it on OwnIdentity for
determining the next edition to insert - it wouldn't be valid for that
purpose.
The next commit will add a getNextEditionToInsert() to OwnIdentity to
avoid that even more.

Also document that it does *not* take edition hints into account.