StatisticsPage: Fix "Sum of all edition numbers" to exclude unfetched editions of seed identities by replacing usage of deprecated Identity.getEdition() with Identity.getLastFetchedEdition()
The point of this statistic is to provide a measurement of how many
editions we have actually fetched, but it wrongly included unfetched
editionts before this commit:

For all non-seed identities which haven't been fetched yet, the value of
getEdition() is equal to what was fetched even if we know a higher
edition hint because edition hints can be fake. So the old code wasn't
wrong for them.
However for seed identities createSeedIdentities() updates them with
known-to-be-good hardcoded latest editions. So their getEdition() may
return an edition which wasn't actually fetched yet so the old code may
have wrongly added non-fetched editions in  their case.