WebOfTrust.restoreOwnIdentity...(): Fix to be robust against user restoring a bogus edition, then deleting the identity, then restoring again with correct edition
... by replacing the deprecated Identity.getEdition() with
getLastFetchedEdition().

The following could happen previously:
- While restoring an identity, the user, for whatever reason, provides
a too high edition which was never actually inserted.
- The user notices that restoring doesn't finish because the edition
didn't exist and thus deletes the identity again.
- The code for deleting an OwnIdentity actually doesn't delete it but
replaces it with a non-own Identity object (to preserve trusts of other
identities to it) *AND* preserves the bogus edition because it cannot
know that it is bogus.
- The hereby replaced code in restoreOwnIdentity...() would copy over
the bogus edition to the newly restored identity by using the deprecated
getEdition() instead of getLastFetchedEdition().

Using getLastFetchedEdition() still is not a great fix: It will only
decrement the bogus edition by one because we don't actually store what
edition was last fetched.
But it is a bit better than before and will benefit from future
improvement of getLastFetchedEdition().