OwnIdentityTest: Fix failure caused by changes to OwnIdentity in previous commit
testGetNextEditionToInsert(plugins.WebOfTrust.OwnIdentityTest) Random seed: -395278536666461957
    FAILED
junit.framework.AssertionFailedError: expected:<Sat Dec 24 20:48:29 CET 2016> but was:<Sat Dec 24 20:48:29 CET 2016>
    at plugins.WebOfTrust.OwnIdentityTest.testGetNextEditionToInsert(OwnIdentityTest.java:170)

- The previous version of this assert didn't really really serve a
the purpose of "testing" expected behavior, I had rather written it to
describe the current behavior as-is.
So I suppose we can do the ugly change of just updating it to describe
the behavior after the changes of the previous commit.

I acknowledge that if WoT was running in multi-user mode, the new
behavior would leak information. But this isn't the only issue of that
type, so it doesn't make things any worse to introduce.
This and the other known issue have been documented in the bugtracker:
https://bugs.freenetproject.org/view.php?id=4840#c12264
https://bugs.freenetproject.org/view.php?id=2554#c12289