WoTTest: Improve readability
- The assert about the edition is moved below the one about the
FetchState because if it fails, that may be because the FetchState is
wrong, so it should be checked after the FetchState.

- The fact that the edition should be equal to the edition hint is just
a side effect which is not interesting here. What matters more is that
the edition hint on the new identity should be the same as on the old
one. So test for that instead.