WoTTest: Resolve (misplaced) FIXME
The FIXME couldn't be resolved where it was at because the Identity
there was NOT an OwnIdentity, it was just a regular one.
I've searched the whole class for calls to
onFetchedAndParsedSuccessfully() and the one which this commit changes
to onInserted() is the only one where it can be replaced.