WoTTest: Replace usage of deprecated Identity.getEdition()
Notice: oldNonOwnIdentity.onFetchedAndParsedSuccessfully(10) had to be
changed to pass something lower than 10 as parameter.
This is because the proper replacement of getEdition() here is
getNextEditionToFetch(), which would return 11 in the case of 10.
But 11 is also the edition we want to pass in the restored insert URI
as next edition to fetch, so if we kept 10 we wouldn't be able to test
for whether 11 is correctly registered.