WoTTest.testDeleteOwnIdentity_Dangling(): Amend to make reason of failure more obvious
Would currently always fail because
Identity.startupDatabaseIntegrityTest() throws because mLastChangedDate
is before mCreationDate - which is because the mCreationDate wasn't
preserved in the replacement non-own Identity.

The failure would happen in the post-test database integrity test, which
doesn't show the underlying exception.
Since initializing the dates improperly could leak information to the
database which could act as proof that the non-own identity used to be
an OwnIdentity it is a good idea to test for this more explicitely
- which this commit thus does.

The failure used to be:

> testDeleteOwnIdentity_Dangling(plugins.WebOfTrust.WoTTest)
> junit.framework.AssertionFailedError
>	at plugins.WebOfTrust.AbstractJUnit3BaseTest.tearDown(AbstractJUnit3BaseTest.java:93)

The failure should now be:

> junit.framework.AssertionFailedError
>   at plugins.WebOfTrust.WoTTest.testDeleteOwnIdentity_Dangling(WoTTest.java:1198)

It will be fixed by an upcoming commit.