WoTTest.testRestoreOwnIdentity...(): Fix to properly test for preventing insertion before restoring is finished
The previous 10 s grace period didn't make any sense because there
is no 10 s sleep between creating the Identity and restoring it so it
couldn't test for mixup of dates.
And even if there was, that wouldn't be thread safe, so we thus now
better check against an actually measured Date.
- Whoever wrote that code must have been drun.. ah damn, it was me :D

(Actually, the previous code did make a bit of sense: It tested against
wrongly initializing the last insert date to the default of new Date(0))

The upcoming commits may fix similar patterns in the other tests, or
maybe I will just file a bug and move on with the actual goals of this
branch :|