OwnIdentity.restoreEdition(): Replace usage of deprecated setEdition(); JavaDoc
Making Identity.forceSetEdition() check whether the edition is lower
than the  minimum allowed edition is necessary because setEdition() used
to implicitely handle this by not doing anything at all if it was.

Throwing is better than doing nothing at all in this case as it will
make the user notice.

Throwing IllegalArgumentException is more appropriate
than throwing InvalidParameterException because it is more standard;
and the caller actually currently fails to handle
InvalidParameterException properly - it doesn't roll back the database
transaction.
A fix for that will be delivered in an upcoming commit.