WebOfTrust.restoreOwnIdentity*(): Fix to rollback database transaction before throwing
Luckily we don't need database upgrade code to fix any potential damage
as the exceptions...
- either happened before anything was written to the database
- or couldn't happen because in the case of e.g. class Trust are thrown
on improper input to the constructors, but we construct the Trust
objects from pre-existing valid other Trust objects.

Nevertheless this should be fixed as it is very likely to cause bugs in
the future if someone adds writes before something which *can* throw.