IdentityTest.testSetEdition(): Replace usage of deprecated setEdition() by recycling it to test onFetchedAndParsedSuccessfully() instead
- It is valid that InvalidParameterException was replaced with
IllegalStateException in the implementation, the only non-test caller
catches that as well by just doing catch(Exception e).

- The test at the end of the function is removed because it isn't
allowed anymore to set an edition as fetched which was previously set
as fetched already. Hence the try/catch before it is adapted to take
account for that / what was missing in it from the removed code block is
added.