Identity.getRawEdition(): JavaDoc
The part about synchronization is removed because it is wrong -
while there is no concurrency in the Identity object which may cause
trouble, accesing the database of objects of type Identity should be
synchronized against the WebOfTrust object.
There's no need to review callers to fix potentially consequences
of the wrong JavaDoc, I've very carefully followed the proper
synchronization pattern for the past years - the removed JavaDoc was
merely a leftover from the very early days of WoT.

The FIXME is downgraded to a TODO because removing the remaining usages
of this function would be too much work for now, it shouldn't be a
release-blocker.