Trust: Provide storage for the edition hint given by the truster
Normally this would be stored at an EditionHint object
- but EditionHint objects only exist while an identity is eligible to be
fetched as they act like a download queue. Once an Identity becomes
distrusted by the community they are deleted.
Thus if the Identity becomes eligible for being downloaded again, we
need a way to re-create the EditionHint objects.
Storing the hints at the Trust objects will serve that purpose.