XMLTransformer: Fix to publish more reliable edition hints
Previously, the published edition hints could be of
FetchState.NotFetched or ParsingFailed, now we gurantee that:
- they are never NotFetched, i.e. hints of an edition which we don't
even know really exists.
- they are not ParsingFailed - as long as the identity subject to the
hint did not publish two or more unparseable XML files in a row.

With regards to the removed comment about potentially many of the hints
on the network being wrong:
This is luckily not the case. That's because when we create an Identity,
we do NOT set the edition of the request URI from a remote hint, but
start at 0. Thus the only way to increment the request URI edition of
an identity is to actually successfully fetch a new one.
An exception to this is the edition 0 - if the request URI has that
edition, it may not actually exist on the network because we didn't
validate it by fetching it before setting it.
Hence the new comment which is added only states that old hints about
the edition 0 may be invalid.