XMLTransformer: Fix to not store an EditionHint of 0 for new identities if truster says he wasn't able to download the first edition yet.
This feature was not necessary:
- In the typical case of new identities which are trusted by more than
one  truster, one of them will for sure have fetched them and published
a positive hint for them.
- In the case of a really new identity, i.e. one which just solved its
first captcha, there is a minimum delay of 1 hour between the truster
discovering the captcha solution and uploading his trust list. That
should be sufficient for the trust to attempt downloading the first
edition of the introduced identity. Once the truster has downloaded it,
the trust list he publishes then will not contain a negative edition
anymore as he fetched the first edition already.