Identity.markForRefetch(): Fix bogus JavaDoc, make more easy to understand
The logic behind when this is called is more complex than what the old
JavaDoc said. See:
- WebOfTrust.updateScoresWithoutCommit()
- WebOfTrust.updateScoresAfterDistrustWithoutCommit()
- WebOfTrust.computeAllScoresWithoutCommit()

Further, it is not interesting here to list the precise conditions when
it is called. So this commit changes the doc to focus on explaining the
motivation behind calling the function.

Also, the function doesn't notify the IdentityFetcher (= nowadays
"IdentityDownloader") but the old JavaDoc did sound like it did - so
this commit fixes the doc to mention that.