IdentityDownloaderSlow: Add getEditionHint(Identity, Identity); trim visibility of getEditionHintByID()
Is more easy to read than usage of getEditionHintByID() since with that
one you have to know what the ID consists of.

Will be needed by a following commit.

(Notice: The previous commit said that this commit here would use
assertDidThrow(), but it doesn't. That isn't a mistake - I've postponed
the commit which will use it. It'll be the same one which uses this new
function here. )