EditionHint: Remove unused deprecated fields
The IDs of the identities have been replaced with direct pointers to
them.
This is necessary because {@link IdentityDownloaderSlow#run()} will need
the target Identity object to be able to obtain its {@link
Identity#getRequestURI()} so it can actually download the URI.

This also will save lots of disk space in the datebase due to not
storing 86 bytes of the ID strings in every EditionHint.
At the worst case the amount of EditionHint objects stored is O(Number
of trusts) = O(N*512) where N is the number of identities.
This is very large, for a typical WoT db there are already > 200 000
Trust objects.
- Thanks to ArneBab for suggesting to reduce disk usage here.

In turn for having introduced storage of Identity objects in
EditionHint, their manager class IdentityDownloaderSlow will have to
ensure that upon deletion of an Identity, the EditionHint objects are
deleted before.
Thus the IdentityDownloader interface's JavaDoc is adapted for that and
a FIXME is added to implement this new requirement.