IdentityDownloaderSlow: Implement run()
Notice: It was intentionally decided to NOT have the callers of
triggerExecution() (which schedules run() to execute) , e.g.
storeNewEditionHintCommandWithoutCommit(), check whether run() will
actually have to start a download or won't do anything because the
number of running downloads is high enough or the new hint is too low on
the queue.
This was decided because it keeps their code simple and run() is a
read-only thread, so in the future once we use ReadWriteLock instead of
synchronized() it will be very fast.
Further downloads can finish in between triggerExecution() and run() so
we couldn't check the number of downloads before calling
triggerExecution() anyway. We could only check the queue position.