IdentityDownloader*: Add constants for network request priority
The amount of EditionHints is O(N*512) where N is the number of
identities on the network.
As N may become very large if we get many users in the long term a
single instance of WoT may not be able to download all of the editions.
Thus we can chose the priority of IdentityDownloaderSlow to be very low.