IdentityDownloaderSlow: Implement getRequestClient(); unify/denoise RequestClient-related documentation in other classes which use a RequestClient.
IdentityFetcher.mRequestClient's JavaDoc, in opposite to the other
changed classes, does not link to getRequestClient() as IdentityFetcher
has no getRequestClient() because it doesn't need to implement a fred
interface which has such a getter. This is due to USK subscriptions in
fred having a different interface architecture than the direct fetches
which the other classes do.

The implementations of getRequestClient() which are overrides of
interface specifications use "@return" instead of specifying the return
value in the JavaDoc body to ensure the interface's JavaDoc body stays
available in the resulting rendered JavaDoc.