IdentityDownloaderSlow: Fix to delete hints given by an Identity once it becomes distrusted
The required synchronization is provided as requirement of the interface
specification of
IdentityDownloader.storeAbortFetchCommandWithoutCommit().

The added comment says that we use getEditionHintsBySourceIdentity() for
handling Identity deletion even though this commit just added the
function without adding that - this is intentional: There is a FIXME at
interface IdentityDownloader to add callbacks for handling Identity
deletion.