IdentityDownloaderSlow: When an Identity becomes distrusted abort running downloads of it, not just queued ones
As opposed to the removed FIXME, it is more wise to implement this here
than on run():
run() is executed very frequently, i.e. possibly for every valid hint of
a trusted identity - but this function here should only be executed once
an Identity becomes distrusted, which hopefully doesn't happen very
often.
So we keep complexity off the function which executes more often.