IdentityDownloaderSlow: Fix wrong logging / remove non-fixworthy FIXME
It's not the job of the callers to manage/understand the hint queue,
they should just give hints to us and it's our job to decide if we need
them. Thus it isn't an error to receive hints which haven't changed, it
just means that an Identity uploaded a new trust list but didn't know
any new edition for it.