StatisticsPage: Add stats about IdentityDownloaderSlow; trim visibility
The functions which are made private aren't called from the outside and
it wouldn't make sense to do so as they only exist for increasing the
readability of make().

Add both English and German l10n since I'm a native speaker of German.
Notice: This isn't uploaded to Transifex yet, will do when I merge the
branch.