IdentityDownloaderSlowStatistics.mSkippedDownloads: Fix to include if a hint is skipped becaue we downloaded a newer hint from the same Identity first
Our download queue sorting algorithm (as specified by
EditionHint.compareTo()) is intentionally designed to download hints
with the newest timestamp first.
In return the downloaded trust lists will have a high probability of
containing the most recent hints for other identities.
Thus the situation of a hint being superseded by a newer one from the
same identity is a direct benefit of our algorithm - and therefore we
are entitled to include it in the metric which judges the quality of our
algorithm, and that is mSkippedDownloads.