IdentityDownloaderSlow.testDatabaseIntegrity(): Move code to EditionHint.startupDatabaseIntegrityTest()
It was wrongly placed at testDatabaseIntegrity() for the reason
explained by the JavaDoc which is added there.

Adding the try/catch block to assume bestCapacity==0 upon
NotInTrustTreeException instead of throwing it out will make that
failure more easy to understand as it will result in the if(bestCapacity
== 0) check throw an IllegalStateException with a related error message
instead of the generic one of NotInTrustTreeException.
Notice: 0 is the smallest legal capacity so doing "bestCapacity == 0"
instead of "< 1" is safe. Also the error message thrown by that if() is
changed to be more descriptive: The previous wording "Identity" was
ambiguous because a hint has two identities.