IdentityDownloaderFast.shouldDownload(): Fix to take trustworthiness into consideration
Yes this looks like a stupid mistake I shouldn't have made - this
happened because I had initially written it for use in instances where
trustworthiness is already given. But I'm not sure whether I will need
it there yet, and meanwhile I'd like to have the function be correct for
all usecases.