IdentityDownloaderFast: Implement DownloadScheduler.run()
mDownloads tracks the String IDs of the Identity objects instead of the
objects themselves because we shouldn't pin objects from db4o in memory
outside of transactions. Their lifecycle should solely be managed by
db4o to ensure the ACID properties of transactions.
For the purpose of managing running downloads we need nothing but their
ID anyway so cloning the Identity objects to be able to keep them in
memory safely w.r.t. db4o would be a waste of RAM.

stopDownloads() is placed before startDownloads() in run() because
freeing memory should be done before consuming more of it.
The function definition of stopDownloads() is however placed after
the one of startDownloads() as for understanding how to free the
resources of a download one should first have read what constitutes
creating one.

The lock on mWoT is taken because we query stuff from it.
The lock mLock is taken to protect mDownloads against concurrent
modification which e.g. may happen once code is added to show the
running downloads on the web interface.