WebOfTrust.computeAllScoresWithoutCommit(): Fix to update all Scores before calling callbacks on IdentityDownloaderSlow
storeStartFetchCommandWithoutCommit() does computations which
potentially involve the Scores of *all* identities so they need to be
updated before we call it.

storeAbortFetchCommandWithoutCommit() doesn't do this but applying the
same change there nevertheless to keep the codebase coherent.

For a minor performance improvement the needStartFetchCommand /
needAbortFetchCommand variables could in theory be
freenet.support.IdentityHashSet instead of IdentifierHashset:
That would use Object.hashCode(), which only has to process the 8 byte
object pointer, instead of using String.hashCode() which has to use
the 43 byte Identity ID strings.
But that would be confusing as "Identity" there means object equality,
it doesn't refer to a WoT Identity.
This improvement can be neglected in favor of the better readability
though:
Since the Score computation improvements of build0018
computeAllScoresWithoutCommit() is not frequently used during normal
operation anymore, it is only used for monthly database coherency
checks.

When filling the needAbortFetchCommand / needStartFetchCommand sets, the
identity is removed from the opposite set because as we process the
Score graphs of multiple OwnIdentitys in the outer loop a single
Identity may have its shouldFetch-state changed multiple times until
we've processed the views of all OwnIdentitys.

When processing the needAbortFetchCommand / needStartFetchCommand
variables, their order is reversed in comparison to their declaration
because:
- the convention of WoT is to declare stuff which involves "starting"
something before the declaration of the symmetrical "stopping" part.
- but when processing those lists, the one for aborting things will free
storage in the database so it is good to do that first to prevent
database fragmentation.

The upcoming commits will also fix this for the other implementations of
Score computation.