WebOfTrust.computeAllScoresWithoutCommit(): Tweak performance
This applies the performance improvement which the previous commit
mentioned and defered in favor of readabiliy.
I'm doing it now because I figured that the necessary documentation to
avoid the potential confusion can be kept sufficiently short by
amending the JavaDoc of class IdentifierHashSet like this commit does
it.

Further when I did the previous commit I failed to notice a more
significant performance issue about the old code - see the added
JavaDoc.