WebOfTrust.updateScoresWithoutCommit(): Fix to update all Scores before calling callbacks on IdentityDownloaderSlow
storeStartFetchCommandWithoutCommit() does computations which
potentially involve the Scores of *all* identities so they need to be
updated before we call it.
storeAbortFetchCommandWithoutCommit() doesn't do this but applying the
same change there nevertheless to keep the codebase coherent.

When filling the needAbortFetchCommand / needStartFetchCommand sets, the
identity is removed from the opposite set because as we process the
Score graphs of multiple OwnIdentitys in the outer loop a single
Identity may have its shouldFetch-state changed multiple times until
we've processed the views of all OwnIdentitys
- albeit I'm less confident about this for this function than I was for
computeAllScoresWithoutCommit() at
93ee8462c0c340a0f9e4c9d52eba14703c1e8904, hence the additional comment
is added in comparison to that.

When processing the needAbortFetchCommand / needStartFetchCommand
variables, their order is reversed in comparison to their declaration
because:
- the convention of WoT is to declare stuff which involves "starting"
something before the declaration of the symmetrical "stopping" part as
to know how to properly stop something it's good to be aware of how it
was started.
- but when processing the lists here, the one for aborting things will
free storage in the database so it is good to do that first to prevent
database fragmentation.

The upcoming commits will also fix this for the other implementations of
Score computation.