WebOfTrust.updateScoresAfterDistrustWithoutCommit(): Fix to deduplicate callbacks on IdentityDownloaderSlow
Normally this commit would have been called "Fix to update all Scores
before calling callbacks on IdentityDownloaderSlow" just like the
previous commit (for the reason of why that is necessary see that
commit's message).
But luckily updateScoresAfterDistrustWithoutCommit() already does this.
Nevertheless a similar code pattern is introduced by this commit:
The way this function here worked before this commit could maybe mean
that multiple calls to  storeAbortFetchCommandWithoutCommit() and/or
storeStartFetchCommandWithoutCommit() could happen for a single Identity
even though only one is necessary.
As the handling of those callbacks is quite complex (and currently not
even safe against duplicate invocation) for both IdentityDownloaderSlow
and IdentityDownloaderFast it is better to only call each handler once.
The code pattern of the previous commit also allows for deduplication
of invocation of the callbacks and thus hereby is also introduced here.

When processing the needAbortFetchCommand / needStartFetchCommand
variables, their order is reversed in comparison to their declaration
because:
- the convention of WoT is to declare stuff which involves "starting"
something before the declaration of the symmetrical "stopping" part as
to know how to properly stop something it's good to be aware of how it
was started.
- but when processing the lists here, the one for aborting things will
free storage in the database so it is good to do that first to prevent
database fragmentation.

The upcoming commits will also fix this for the remaining callers of
storeStartFetchCommandWithoutCommit() and
storeAbortFetchCommandWithoutCommit().