WebOfTrust.restoreOwnIdentity...(): Fix to update Scores before calling callbacks on IdentityDownloaderSlow
storeStartFetchCommandWithoutCommit() does computations which
potentially involve the Scores of all identities so they need to be
updated before we call it.
storeAbortFetchCommandWithoutCommit() doesn't do this but the previous
commits introduced the convention to also only call it when the Score
database is up to date to keep the codebase coherent.
This will in a following commit be documented in the JavaDoc of the
IdentityDownloader interface once all callers are changed to obey it.

The permission to store references to Identity objects has already been
added to the IdentityDownloader JavaDoc by a previous commit.
However for the new implementations deletion of obsolete references
won't be handled by triggering command processing as the removed comment
indidicated but instead in the store...Command() callbacks right away.