IdentityDownloaderFast: Make download scheduling thread robust against unexpected errors; backport documentation to IdentityDownloaderSlow
The delay of 1 minute is necessary to avoid potential CPU starvation
upon errors which happen at every iteration.
Not using a constant as in opposite to IdentityDownloaderSlow the
IdentityDownloaderFast will *not* have a batching delay:
Its only interest is identities to which the user has manually assigned
a trust value and the user's actions should be processed immediately
instead of having a delay.