IdentityFetcher: Finally figure out why a commented-out assert used to fail and document that.
Not removing the assert/the TODO above it as the whole class is supposed
to be replaced by this branch and thus not much work should be invested
into it.
Still this needs to be documented a bit now by this commit as there is a
bugtracker entry related to it - once we want to deal with that entry
this comment will help a lot.