IdentityDownloaderFast: Add member class DownloadSchedulerCommand
Based on IdentityFetcher.IdentityFetcherCommand.
Like there it is intended as base class for upcoming child classes which
indicate whether a download shall be started/stopped.

Instead of IdentityFetcherCommand, it stores pointers to the Identity
objects themselves instead of merely a String copy of their ID:
When starting downloads we will need to query the URIs from the
Identity objects anyway. Storing the ID only would require another
database query to obtain the Identity object.
It's not a big problem to store the Identity object with regards to the
WebOfTrust deleting Identity objects and thus zeroing the pointers:
This will have to be dealt with anyway due to the IdentityDonloader
interface recently having been changed to allow implementations to store
pointers. See the amended FIXME at WebOfTrust.

Instead of being called "IdentityDownloaderFastCommand" it is called
DownloadSchedulerCommand because it will be the input to the member
class DownloadScheduler which IdentityFetcher did not have.

The pointers which are passed to the constructor are only validated with
an assert() as the constructor is only visible to internal code, not to
things which pass user input.