IdentityDownloaderFast: Add getQueuedCommand()
Will be used by storeStartFetchCommandWithoutCommit() to check whether
a command already exists.
Hence the word "Queued" is included in the name to ensure it doesn't
sound like this function is used to obtain a new command object to
store.

In opposite to most of WoT, when the database query yields no result,
null is returned instead of throwing an exception as that would be
slower.
This optimization is worth it because yielding no result is the normal
expected mode of operation when being called from
storeStartFetchCommand...():
That function is typically called for a newly introduced identity, and
for new identities there won't be existing commands.
Commands will only preexist if e.g. the "do we want to fetch it?" state
changes quickly from true to false *and* to true again, which probably
is a rather rare event.