IdentityDownloaderFast: Finish fix to obey Identity.markForRetch()
The series of previous commits about this is completed now.

Based on IdentityFetcher.fetch(Identity).

Logging priority is raised from minor to normal as this mechanism is
pretty fragile (due to the lack of an explicit "refetch requested" flag
and instead guessing it from the FetchState) and thus we should have a
look at whether it does work properly.