IdentityDownloaderFast.onFound(): Implement
Based on IdentityFetcher.onFound() and
IdentityDownloaderSlow.onSuccess() (which itself is an improvement
version of the former).

The try{} block is expanded to contain the whole function due to a bug
in the base functions:
The functions were written under the assumption that
FetchResult.asBucket() creates the Bucket as it is called, i.e. the
Bucket doesn't exist before it has been called and thus doesn't have to
be freed if the function isn't called.
But this isn't the case - FetchResult objects contain a reference to the
Bucket right from their creation on.

This has been documented in the bugtracker so it will be fixed for the
base function at a later point in time:
https://bugs.freenetproject.org/view.php?id=6947

As opposed to IdentityDownloaderSlow.onSuccess() this doesn't
delete edition hints, remove the download from the mDownloads map or
trigger the download scheduler thread for execution:
IdentityDownloaderFast does USK subscriptions instead of hint based SSK
fetches, and USK subscriptions are meant to be permanent.
(It would make sense to notify the IdentityDownloaderSlow that it should
delete its hint, but a FIXME to implement this already is present at
that class' JavaDoc.)

I'm still using Closer.close() instead of the try-with-resources
statements of Java 7 as IIRC Bombe said on IRC that some of fred's
classes aren't comfortable with it yet.
So converting WoT to use that will be done in bulk to ensure proper
testing - as part of the bugtracker entry for that:
https://bugs.freenetproject.org/view.php?id=6949