IdentityDownloaderFast.storeStartFetchCommand...(): Introduce special version of this function for the purposes of storeTrustChangedCommandWithoutCommit()
As demanded by the FIXME there (which will be removed once the similar
aspect of introducing a special storeABORT...() is resolved).

The new function doesn't have to handle markForRefetch() as the WoT core
will call storeStartFetchCommand...() for that purpose already so
storeTrustChangeCommand...() doesn't have to detect whether it was
requested.