WebOfTrust.removeTrust...(): Call IdentityDownloader.storeTrustChanged...()
As demanded by its specification.

Will maybe file a bug about documenting that the callback, and also
SubscriptionManager's callback, doesn't receive a clone() under this
circumstance, despite the fact that the documentation of these callbacks
says they receive clones.
It's not a critical issue IMHO, the primary reason why they expect
clones is because if the Trust isn't removed but only changed then you
obviously cannot pass the old state of the Trust with the same Trust
object. I.e. they don't want a clone() because they would e.g. plan to
tamper with the Trust's Identity objects which would cause damage if it
weren't a clone. So it doesn't hurt if it isn't a clone.