AbstractFullNodeTest: Rename to AbstractMultiNodeTest
Intentionally NOT adapting any of the classes which use it to reflect
the new name:
The goal of this series of commits is to introduce AbstractMultiNodeTest
as a new base class for class AbstractFullNodeTest while preserving Git
history of the code by renaming the original class to the name of the
base class.
The following commit will thus add a new almost empty class
AbstractFullNodeTest as a subclass of what has become
AbstractMultiNodeTest now.