AbstractJUnit4BaseTest: Remove comments which were wrongly copy-pasted from AbstractMultiNodeTest
They don't apply because AbstractJUnit4BaseTest doesn't run WoT inside a
Freenet node, it just runs the WebOfTrust class directly.

Notice: This commit doesn't add the commented-out assert which the
removed comment talks about because in fact it was already added to the
code when it was copy-pasted, see the lines above the removed comment.

Commit e07217c7d6c7f58cb80f2d9ac037abd3b9e27304 is at blame for this.

(Fixing it on this branch even though it doesn't belong here as it is a
trivial change and I just spotted it while reviewing
AbstractMultiNodeTest for the purposes of this branch.)