AbstractMultiNodeTest.printNodeStatistics(): Extract functions to compute each statistical value
To try to decrease the overall runtime of the tests I will try to use
this during connectNodes() to wait for the network to settle to
reasonable backoff/ping times before doing the actual tests:
I would guess that fred has code to wait until peers aren't overloaded
before it executes inserts/requests, and those waits may be longer than
it actually takes for the network to settle.
I.e. it may act as "if(overloadedWhenStartingTheInsert)
waitMultipleMinutes()" which is too much waiting.
- Notice that I'm just blindguessing here though, this may not actually
be the case and I may thus not commit such changes. Nevertheless having
extracted functions for this may be useful in the future.

The comment about the executor is changed to refer to it by the name of
the member variable which holds it to ensure if someone wants to change
the said fact about the executor they will run into the comment and
notice they have to adapt the code there as well.