Cherry-pick from next: Travis CI: Fix fred to compile
by using the new Gradle builder instead of the removed Ant one.

This is a squashed cherry-pick of the commits starting at
	40e77e933cee0d90d30372586912e6f6f0c81d02
up to
	00364b2aec3bf9347d096fa5ecd82e2f9779bbc9
of branch next.

I.e. it is the changes of branch issue-0006929-fix-travis except for the
last commit which amends the changelog. That commit will follow this one
as a single cherry-pick.

The commits have been squashed because they will be merged along
with branch next after this branch anyway so the full history can
already be seen at the original commits then.

We have to cherry-pick their changes instead of merging them
because branch next contains preceding changes which are not ready for
release yet, but we need to release build0020 soon in order to make WoT
compatible with the upcoming removal of the db4o library from fred's
freenet-ext.jar.

Fixes https://bugs.freenetproject.org/view.php?id=6929
= "Travis CI builds fail due to changes at fred"