SubscriptionManagerFCPTest: Documentation
Describing the reason for this behavior is better than just describing
the consequences of it because it makes it more easy to locate the code
which causes it in case someone needs to get rid of the behavior
someday.

Also the old documentation wasn't correct because SubscriptionManager
doesn't directly use a Ticker anyway, it uses a
TickerDelayedBackgroundJob.