AbstractMultiNodeTest: Simplify / fix documentation
The sentence about taking the locks while doing getAllIdentities() /
getAllTrusts() / getAllScores() only for coherency was wrong - those
functions return ObjectSets so concurrent modification of the database
could be a problem due to
https://bugs.freenetproject.org/view.php?id=6247
- see the attached RollbackFailFixed.java: Pinning things into memory
with the current way WoT does transactions will break rollback().
And even if this weren't the case it would be apparent enough that it is
for coherency so we don't need to document that.