Logging: Add at IntroductionClientTest, improve at IdentityFileProcessor & IdentityFetcherTest
Primary goal is to set up logging for IntroductionClientTest.
For that purpose also improving the logging at IdentityFileProcessor and
backporting the improvements of getDetailedLogLevel() to
IdentityFetcherTest whose version of that function inspired the one
hereby added.

The logging added to IdentityFileProcessor is MINOR instead of NORMAL
because it causes the additional computation of converting the time to a
string, and MINOR instead of DEBUG because the additional computation
does not change behavior / is not heavy.