IntroductionPuzzleStore: Add getInsertedOwnPuzzlesByInserter(); fix documentation
This function was lacking in the previous commit, sorry.

The documentation of synchronization requirements at
getUninsertedOwnPuzzlesByInserter() was wrong.
It is replaced by documenting synchronization requirements at the class
instead. Documenting of synchronization requirements at every function
would be a waste of time before
https://bugs.freenetproject.org/view.php?id=6247 is resolved as that
will also change the whole synchronization requirements (e.g. by
introducing ReadWriteLocks).
I've reviewed the existing callers of
getUninsertedOwnPuzzlesByInserter(), they do obey the requirements.

The documentation of what the function does was trivial and is hence
removed.
The documentation about what uses it was also trivial, you can
just look that up with an IDE.

Further the visbility can be trimmed hereby.