IntroductionClientTest: Fix race condition & lack of proper synchronization
Due to redundancy the amount of data to upload is usually larger than
what has to be downloaded, so fred's "upload finished!" callbacks can
happen AFTER the remote node's "download finished!" callbacks have
already returned.
This means that the loop which waits for puzzle up-/download can break
before the upload was finished and hence the puzzleID/puzzleSolution
variables could have been left uninitialized.

As a consequence of initializing them before the loop we don't need
IntroductionPuzzleStore.getInsertedOwnPuzzlesByInserter() anymore, which
was added by commit 3aa44f833b63788fc147706fddbbcde741d70d7d solely for
our purposes here, so it is hereby removed again.

The lack of synchronization was that getOwnCatpchaAmount() requires
being synchronized on the WebOfTrust and IntroductionPuzzleStore (as
specified by the class-level JavaDoc of IntroductionPuzzleStore), but
only takes the lock on the IntroductionPuzzleStore on its own, so we
need to take the lock on the WebOfTrust.